feat: add dry-run option (#34)
close #32

+++ src/main.ts
@@ -2,22 +2,31 @@ import * as core from '@actions/core'
 import * as github from '@actions/github'
 import { getRef } from './ref'
 import type * as types from '@octokit/openapi-types'
+import { getInputs } from './get-inputs'
 
 type Cache =
   types.components['schemas']['actions-cache-list']['actions_caches'][number]
 
+const ansi = { reset: '\x1B[0m', dryRun: '\x1B[38;2;90;185;255m' }
+
+const prefix = ({ isDryRun = false }: { isDryRun: boolean }): string =>
+  isDryRun ? `${ansi.dryRun}DRY-RUN MODE ${ansi.reset}` : ''
+
 const deleteRefActionsCaches = async (
   octokit: ReturnType<typeof github.getOctokit>,
   repo: { owner: string; repo: string },
-  ref: string
+  ref: string,
+  isDryRun: boolean
 ): Promise<void> => {
   const deleteCache = async (cache: Cache): Promise<void> => {
     if (!cache.id) return
-    core.info(`   - Cache with key ${cache.key}`)
-    await octokit.rest.actions.deleteActionsCacheById({
-      ...repo,
-      cache_id: cache.id
-    })
+    core.info(`${prefix({ isDryRun })}   - Cache with key ${cache.key}`)
+    if (!isDryRun) {
+      await octokit.rest.actions.deleteActionsCacheById({
+        ...repo,
+        cache_id: cache.id
+      })
+    }
   }
 
   // https://github.com/octokit/plugin-paginate-rest.js#octokitpaginate
@@ -29,7 +38,9 @@ const deleteRefActionsCaches = async (
       per_page: 100
     }
   )
-  core.info(`⌛ Deleting ${caches.length} cache(s) on ${ref}`)
+  core.info(
+    `${prefix({ isDryRun })}⌛ Deleting ${caches.length} cache(s) on ${ref}`
+  )
 
   await Promise.all(caches.map(async cache => deleteCache(cache)))
 }
@@ -40,14 +51,13 @@ const deleteRefActionsCaches = async (
  */
 export async function run(): Promise<void> {
   try {
-    const token = core.getInput('repo-token', { required: true })
+    const { token, dryRun: isDryRun } = getInputs()
     const octokit = github.getOctokit(token)
 
     // get repostiory information
     const { repo, eventName, payload } = github.context
 
     const ref = getRef({ eventName, payload })
-
     if (ref === null) {
       core.info('🤔 Could not determine deletion target.')
       core.info(
@@ -55,8 +65,9 @@ export async function run(): Promise<void> {
       )
       return
     }
-    await deleteRefActionsCaches(octokit, repo, ref)
-    core.info('✅ Done')
+    await deleteRefActionsCaches(octokit, repo, ref, isDryRun)
+
+    core.info(`${prefix({ isDryRun })}✅ Done`)
   } catch (error) {
     // Fail the workflow run if an error occurs
     if (error instanceof Error) {
